+++ lib/plug/accesslog/formatter/query_string.ex
@@ -1,0 +1,16 @@
+defmodule Plug.AccessLog.Formatter.QueryString do
+  @moduledoc """
+  Logs the query string.
+  """
+
+  @doc """
+  Appends to log output.
+  """
+  @spec append(String.t, Plug.Conn.t) :: String.t
+  def append(message, conn) do
+    case conn.query_string do
+      ""    -> message
+      query -> message <> "?" <> query
+    end
+  end
+end