+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Additional formatting directives
   - `%a` - Remote IP-address
+  - `%q` - Query string (prepended with "?" or empty string)
   - `%T` - Time taken to serve the request (full seconds)
   - `%V` - Server name (canonical)